tests: Updated golden standards on Windows.

+++ yumo/app.py
@@ -183,7 +183,7 @@ class PolyscopeApp:
 
         psim.Separator()
 
-    def _handle_mesh_pick(self, pick_result: ps.PickResult) -> float | None:  # type: ignore[no-any-unimported]
+    def _handle_mesh_pick(self, pick_result: ps.PickResult) -> np.float64 | None:  # type: ignore[no-any-unimported]
         """Handle mesh picking cases: face, vertex, corner."""
         logger.debug("Picked mesh")
         mesh: MeshStructure = self.structures["mesh"]  # type: ignore[assignment]
